// frontend/src/types/student/profile.ts
export const additionalDetailSchema = z.object({
    title: z.string().min(1, "Title is required"),
    description: z.array(z.string()).min(1, "Description is required"),
    date: dateString.optional(),
});
// ...
export type AdditionalDetailInput = z.infer<typeof additionalDetailSchema>;
export type UpdateAdditionalDetailInput = Partial<AdditionalDetailInput>;
export type AdditionalDetail = AdditionalDetailInput & { id: number; profileId: number };